+++ src/component/unit/answer/write/answerwrite.styles.ts
@@ -17,6 +17,7 @@ export const Person = styled.img`
 `;
 export const CommentBox = styled.div`
   width: 90%;
+  margin-left: 15px;
 `;
 export const Top = styled.div`
   /* background-color: red; */
