+++ src/component/unit/landing/landing.container.tsx
@@ -1,10 +1,22 @@
-import { MouseEvent, useRef, useState } from "react";
+import { useRouter } from "next/router";
 import LandingPresenter from "./landing.presenter";
 
 export default function LandingContainer() {
+  const router = useRouter();
+
+  const onClickMoveToList = () => {
+    router.push("/boards");
+  };
+  const onClickMoveBoardNew = () => {
+    router.push("/boards/new");
+  };
+
   return (
     <>
-      <LandingPresenter />
+      <LandingPresenter
+        onClickMoveToList={onClickMoveToList}
+        onClickMoveBoardNew={onClickMoveBoardNew}
+      />
     </>
   );
 }
