+++ src/component/unit/landing/landing.types.ts
@@ -1,3 +1,4 @@
-import { MouseEvent } from "react";
-
-export interface ILandingPresenter {}
+export interface ILandingPresenter {
+  onClickMoveToList: () => void;
+  onClickMoveBoardNew: () => void;
+}
